removed unintentional implicit int

+++ tahoe/src/elements/continuum/solid/materials/Simo-Holzapfel/FDSimoVisco2D.cpp
@@ -1,4 +1,4 @@
-/* $Id: FDSimoVisco2D.cpp,v 1.2 2002-10-05 03:44:31 paklein Exp $ */
+/* $Id: FDSimoVisco2D.cpp,v 1.3 2002-10-14 16:04:06 thao Exp $ */
 /* created:   TDN (5/31/2001) */
 
 #include "FDSimoVisco2D.h"
@@ -23,8 +23,8 @@ FDSimoVisco2D::FDSimoVisco2D(ifstreamT& in, const FiniteStrainT& element):
 	in >> ftauS;
 	in >> ftauB;
 
-	int taudtS = fdt/ftauS;
-	int taudtB = fdt/ftauB;
+	double taudtS = fdt/ftauS;
+	double taudtB = fdt/ftauB;
 
 	falphaS = exp(-0.5*taudtS);
 	falphaB = exp(-0.5*taudtB);
